make todo component less dependent
previously, the todo component required a lot of information about its
parent so that it could supply the right information to its callbacks.
now the callbacks only provide the minimum information required, so the
component doesn't need as much information as before. this means that it
will be easier to use in a wider variety of situations.