Handle values provided by AutoFill or other means
This adds a polling-based solution to the issue where React can't see values provided by browser AutoFill features and add-ons.
If the value of the `autoComplete` property on an `InputText` component is not `"off"`, any values provided by means other than normal user input (`onChange`) has additional handling. This catches both automatic and user-initiated form filling.

This has the unfortunate side effect of making `valueLink` unusable as React will complain about using both it and `onChange`, which we need to use in order to elegantly handle managed and non-managed value changes.

part of a solution for macropodhq/pouch-client#75